+++ R/r2_adjusted.R
@@ -1,4 +1,4 @@
-#' Calculate the adjusted determination coefficient
+#' Calculate the Adjusted Determination Coefficient
 #' @description
 #' Calculate the adjusted coefficient of determination by entering the regression model and coefficient of determination.
 #' See details.
@@ -31,9 +31,10 @@
 r2_adjusted <- function(model, r2){
   v <- values_lm(model)
 
-  r2_adjusted <- 1 - (1 - r2) * v$a
-  names(r2_adjusted) <- paste(names(r2), "adjusted", sep = "_")
-  class(r2_adjusted) <- "r2_kvr2"
+  r2_adj <- 1 - (1 - r2) * v$a
+  names(r2_adj) <- paste(names(r2), "adj", sep = "_")
 
-  r2_adjusted
+  # r2_adj <- set_kvr2_attr(r2_adj, v = v, class_name = "r2_kvr2")
+
+  r2_adj
 }